backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
pickle-mixin==1.0.2
matplotlib==3.7.2
seaborn==0.12.2
joblib==1.3.2
Pillow==10.0.0
transformers==4.33.2
torch==2.0.1
bcrypt==4.0.1
PyJWT==2.8.0